pyproject.toml
[project]
name = "famp"
version = "0.1.0"
description = "Facebook Account Management Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nodriver",
    "click",
    "pydantic",
    "tabulate>=0.9.0",
    "cryptography>=44.0.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort"
]

[project.scripts]
famp = "famp.cli:main"
